Added CloudValidator object.
The CloudValidator is used to get the sender of whoever set a cloud variable. Its method, validate_cloud, takes a CloudChange object and sets the sender attribute using the clouddata logs. More tests are needed to see how reliable the current system is, but to my knowledge, nobody has a validator in their wrapper yet :)