Created Documentation
This commit adds documentation pulled from the docstrings, a basic documentation website, and new pages.

All documentation has been created using sphinx.

basic_tutorial is a basic tutorial on scratchcloud that explains how its event system is set up.

getting_started is a basic intro to cloud variables and it has install info.

index is the homepage.

examples is empty for now.

A built version will be added in the next commit.